Bring graphsync specs back to life.
This is long belated.  (The original reason for stalling on them was
that I hadn't quite figured out, ontologically speaking, where they
seemed to best belong.  That question was actually answered when we
had discussions about where to put the CAR specifications: the answer
for that is clearly also the correct answer for where to put Graphsync.
So, without further ado: let it be so.)

Also added some intro to the transport index page; some light synopsis;
and the minimal essential link fixing.