adls: signalling: add additional examples.
The key point is that there are **many** places where signalling
can show up.  None is the clear single mechanism that beats others.

This is just quick bullet points for now.  It should be expanded.
I'm having a very "not enough hours in the day" day, so it will
have to come later.