dag-cbor,dag-json: relax strictness rules for decoders (#214)
* dag-cbor,dag-json: relax strictness rules for decoders

Fixes: https://github.com/ipld/ipld/issues/196

Current text reflects an idealistic perspective, not a pragmatic one that
accounts for historical data and handling data from encoders that don't quite
follow the rules. Nor do the current statements requiring strictness reflect
the reality of our current decoders, which are lose and don't even yet have
full opt-in strictness modes; even though we note the lack of strictness in
the implementation notes.